requirements.txt
PySide6>=6.6,<6.9
pyserial>=3.5
numpy>=1.24
pydub>=0.25
